requirements.txt
streamlit==0.90.0
pytesseract==0.3.8
pillow==8.3.2
